backend/requirements.txt
djangorestframework==3.11.2
Django==3.1.8
django-tinymce
django-cors-headers
django-cleanup
psycopg2-binary==2.9.1
gunicorn==20.1.0
Pillow
whitenoise